refactor : MyLottos클래스 분리 MyLottos 클래스의 필드를 컨트롤러에 의존하여 수정하기 때문에 MyLottos에서 결과 값과 수익률을 따로 모델로 만들어서 분리
